refactor(components) : group modal components
