+++ src/app/auth/verify-user/page.tsx
@@ -1,7 +1,7 @@
 import { SignInForm } from "../../../components/auth/SignInForm";
 import { generateSecurityTokenAction } from "../../../server-action/backend/auth/api";
 import { OneTimeTokenPurposeValues } from "../../../server-action/backend/auth/type";
-import ErrorModal from "../../../components/ErrorModal";
+import ErrorModal from "../../../components/modal/ErrorModal";
 import AuthFooter from "../../../components/auth/AuthFooter";
 import { AUTH_LOGIC_ROUTE } from "../../../route/auth/route";
