+++ src/app/quiz/[id]/edit/page.tsx
@@ -1,4 +1,4 @@
-import ErrorModal from "../../../../components/ErrorModal";
+import ErrorModal from "../../../../components/modal/ErrorModal";
 import QuizForm from "../../../../components/quiz/QuizForm";
 import { getSignInInfo } from "../../../../server-action/backend/cookie";
 import { getQuizAction } from "../../../../server-action/backend/quiz/api";
