+++ src/app/quiz/create/page.tsx
@@ -1,4 +1,4 @@
-import ErrorModal from "../../../components/ErrorModal";
+import ErrorModal from "../../../components/modal/ErrorModal";
 import QuizForm from "../../../components/quiz/QuizForm";
 import { getSignInInfo } from "../../../server-action/backend/cookie";
