+++ src/components/auth/deleteAccountForm.tsx
@@ -5,9 +5,9 @@ import { deleteUserAction } from '../../server-action/backend/user/api';
 import toast from 'react-hot-toast';
 import { isHttpException, isServerActionError } from '../../server-action/backend/util';
 import { HttpStatus } from '../../server-action/backend/status';
-import GuideModal from '../GuideModal';
+import GuideModal from '../modal/GuideModal';
 import { useRouter } from 'next/navigation';
-import ErrorModal from '../ErrorModal';
+import ErrorModal from '../modal/ErrorModal';
 
 interface DeleteAccountState {
     error : string;
