+++ src/components/generate/CuratedArtworkSection/CuratedArtworkForm.tsx
@@ -1,6 +1,6 @@
 import React, { useReducer } from 'react';
 import CustomInput from '../../CustomInput';
-import DropdownMenu from '../../DropdownMenu';
+import DropdownMenu from '../../modal/DropdownMenu';
 import { FormState, getEmptyFormState } from './states';
 import { _CuratedContentType } from '@/model/interface/curatedArtwork-types';
